move utils inside lambda function
