+++ source/aws-bootstrap-kit/lib/dns/delegation-record-handler/utils.ts
